+++ bee-network/bee-gossip/src/network/host.rs
@@ -112,11 +112,11 @@ async fn network_host_processor(
     loop {
         tokio::select! {
             _ = &mut shutdown => break,
-            event = (&mut swarm).next() => {
+            event = swarm.next() => {
                 let event = event.ok_or(crate::Error::HostEventLoopError)?;
                 process_swarm_event(event, &internal_event_sender, &peerlist).await;
             }
-            command = (&mut internal_command_receiver).recv() => {
+            command = internal_command_receiver.recv() => {
                 let command = command.ok_or(crate::Error::HostEventLoopError)?;
                 process_internal_command(command, &mut swarm, &peerlist).await;
             },