+++ bee-pow/src/providers/miner.rs
@@ -36,7 +36,7 @@ pub enum Error {
     #[error("the worker has been cancelled")]
     Cancelled,
     /// Invalid proof of work score.
-    #[error("invalid proof of work score {0}, requiring {} trailing zeros")]
+    #[error("invalid proof of work score {0}, requiring {1} trailing zeros")]
     InvalidPowScore(f64, usize),
 }