+++ bee-storage/bee-storage-sled/src/access/iter.rs
@@ -58,14 +58,6 @@ macro_rules! impl_iter {
                         .map(|(key, value)| Self::unpack_key_value(&key, &value))
                         .map_err(From::from)
                 })
-
-                // inner.status()?;
-                //
-                // if inner.valid() {
-                //     Poll::Ready(item)
-                // } else {
-                //     Poll::Ready(None)
-                // }
             }
         }
     };