+++ bee-network/bee-gossip/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased - 2022-XX-XX
 
+### Changed
+
+- Updated dependencies;
+
 ### Fixed
 
 - Clippy warnings;