+++ bee-network/bee-gossip/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## Unreleased - 2022-XX-XX
+
+### Changed
+
+- Updated dependencies;
+
+### Fixed
+
+- Clippy warnings;
+
 ## 1.0.0-beta.1 - 2022-07-20
 
 First beta release.