+++ bee-network/bee-gossip/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## Unreleased - 2022-XX-XX
+## 1.0.0-beta.2 - 2022-08-15
 
 ### Changed